+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Updated Mysten gRPC protobuffs
+
 ### Removed
 
 ## [0.94.0] - 2025-12-11